add header nav to whoweare, fixup i18n issues
