+++ public/getCountry.js
@@ -1,5 +1,5 @@
-// TODO(ajwong): This is copied into donation-form-bounce.html. Careful.
-export default () => {
+// TODO(ajwong): This is copied into donation-form-bounce.handlebars. Careful.
+export function getCountry() {
   const url = new URL(window.location);
   const directories = url.pathname.split("/");
   if (directories.length > 2) {
