+++ public/locations-list-map.js
@@ -1,9 +1,10 @@
 import toDataByLocation from './toDataByLocation.js';
 import countries from './countries.js';
 import locales from './locales.js';
-import getCountry from './getCountry.js';
+import {getCountry} from './getCountry.js';
 import { FILTER_ITEMS, ENUM_MAPPINGS } from './formEnumLookups.js';
 import { getMapsLanguageRegion, getCurrentLocaleParam, DEFAULT_LOCALE } from  './i18nUtils.js';
+import {ac, ce, ctn} from './utils.js';
 
 
 /******************************************
@@ -54,74 +55,6 @@ let lastLocationRendered = -1;
 /*************************
  * END MODULE LEVEL VARS *
  *************************/
-
-const generateBottomNav = () => {
-  const currentLocale = getCurrentLocaleParam(DEFAULT_LOCALE);
-
-  const localeDropdownLink = document.getElementById('locales-dropdown');
-  const countryDropdownLink = document.getElementById('countries-dropdown');
-  const localeDropdownItems = document.getElementById('locales-dropdown-selector');
-  const countryDropdownItems = document.getElementById('countries-dropdown-selector');
-
-  if (localeDropdownLink && countryDropdownLink && localeDropdownItems && countryDropdownItems) {
-    const sortedLocales = locales.sort((localeA, localeB) => {
-      const aLocalized = $.i18n(localeA.i18nString);
-      const bLocalized = $.i18n(localeB.i18nString);
-      return aLocalized.localeCompare(bLocalized);
-    });
-
-    sortedLocales.forEach((locale) => {
-      if (locale.localeCode.toLowerCase() === currentLocale.toLowerCase()) {
-        localeDropdownLink.textContent = $.i18n(locale.i18nString);
-      }
-
-      const element = document.createElement('a');
-      element.className = 'dropdown-item';
-      const currentUrl = new URL(window.location.href);
-      currentUrl.searchParams.set('locale', locale.localeCode);
-      element.setAttribute('href', currentUrl.href);
-      element.textContent = $.i18n(locale.i18nString);
-      element.addEventListener("click", () =>  sendEvent("i18n", 'set-locale', locale.localeCode));
-      localeDropdownItems.appendChild(element);
-    });
-
-    const sortedCountryKeys = Object.keys(countries).sort((a, b) => {
-      const countryA = countries[a];
-      const countryB = countries[b];
-
-      const aLocalized = $.i18n(countryA.i18nString);
-      const bLocalized = $.i18n(countryB.i18nString);
-
-      return aLocalized.localeCompare(bLocalized);
-    });
-
-    sortedCountryKeys.forEach((countryCode) => {
-      const country = countries[countryCode];
-
-      if (country.countryCode === currentCountry.toLowerCase()) {
-        const img = ce('div', `icon icon-cf_${country.countryCode}`);
-        ac(countryDropdownLink, [img, ctn($.i18n(country.i18nString))]);
-      }
-
-      const element = document.createElement('a');
-      element.className = 'dropdown-item i18n';
-      const currentUrl = new URL(window.location.href);
-      const pathname = currentUrl.pathname;
-      const updatedPath = pathname.replace(/(\/[a-z]{2}\/|\/)/, `/${country.countryCode}/`);
-      currentUrl.pathname = updatedPath;
-
-      element.setAttribute(
-        'href',
-        currentUrl.href
-      );
-
-      const img = ce('div', `icon icon-cf_${country.countryCode}`);
-      ac(element, [img, ctn($.i18n(country.i18nString))]);
-      element.addEventListener("click", () =>  sendEvent("i18n", 'set-country', country.countryCode));
-      countryDropdownItems.appendChild(element);
-    });
-  }
-};
 
 // i18n must be loaded before filter items can be translated
 // config stores the i18n string and this function calls i18n with it
@@ -255,33 +188,6 @@ function createFilterElements(data, filters) {
   return filterElements;
 }
 
-// Wrapper for Node.appendChild.
-// child (either a node or an array of nodes) is appended to the parent element
-function ac(el, child) {
-  if (el && child) {
-    if (Array.isArray(child)) {
-      child.forEach((c) => el.appendChild(c));
-    } else {
-      el.appendChild(child);
-    }
-  }
-}
-
-// Wrapper for document.createElement - creates an element of type elementName
-// if className is passed, assigns class attribute
-// if child (either a node or an array of nodes) is passed, appends to created element.
-function ce(elementName, className, child) {
-  const el = document.createElement(elementName);
-  className && (el.className = className);
-  child && ac(el, child);
-  return el;
-}
-
-// Wrapper for document.createTextNode
-function ctn(text) {
-  return document.createTextNode(text);
-}
-
 // Turns a string with embedded \n characters into an Array of text nodes separated by <br>
 function multilineStringToNodes(input) {
   const textNodes = input.split('\n').map((s) => ctn(s));
@@ -373,8 +279,6 @@ $(function () {
 
   // this should happen after the translations load
   $('html').on('i18n:ready', function () {
-    generateBottomNav();
-
     $('.add-donation-site-form')
       .attr({href: `/${ currentCountry }/donation-form?locale=${$.i18n().locale}`})
       .click(function(e) {
