+++ views/layouts/static.handlebars
@@ -16,7 +16,8 @@
   <title>#Findthemasks</title>
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <link rel="stylesheet" type="text/css" href="../public/style.css">
+  <link rel="stylesheet" type="text/css" href="style.css">
+  <link rel="stylesheet" href="images/icons.data.svg.css">
   <link href="https://fonts.googleapis.com/css?family=Roboto:400,700&display=swap" rel="stylesheet">
   <meta name="description" content="Help #findthemasks for covid-19" />
   <meta name="keywords" content="#findthemasks, masks, n95, coronavirus, covid-19" />
@@ -35,7 +36,7 @@
   <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.i18n/1.0.7/jquery.i18n.parser.min.js"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.i18n/1.0.7/jquery.i18n.emitter.bidi.min.js"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.i18n/1.0.7/jquery.i18n.emitter.min.js"></script>
-  <script type="module" src="../public/i18n.js"></script>
+  <script type="module" src="i18n.js"></script>
 
   <link rel="canonical" href="https://findthemasks.com/whoweare.html" />
