+++ views/partials/top_nav.handlebars
@@ -1,10 +1,11 @@
+<script type="module" src="render-nav.js"></script>
 <nav class="navbar sticky-top navbar-expand-lg navbar-top navbar-dark ftm">
   <ul class="navbar-nav navbar-brand">
     <li class="nav-item">
       <div class="icon icon-ftm_logo" />
     </li>
     <li class="nav-item">
-      <span class="text">Find The Masks</span>
+      <a class="home-link" onclick="window.location='index.html'+window.location.search;"><span class="text">Find The Masks</span></a>
     </li>
   </ul>
   <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarCollapse" aria-controls="navbarCollapse" aria-expanded="false">
