+++ views/whoweare.handlebars
@@ -1,3 +1,4 @@
+{{> top_nav }}
 <main class="content container py-5">
   <h1>#findthemasks</h1>
